Hoist the serial port / baud into the caller
