+++ client.py
@@ -1,4 +1,6 @@
-from uwh.comms import XBeeServer, XBeeClient
+from uwh.xbee_comms import XBeeServer, XBeeClient
+from uwh.gamemanager import GameManager
 
-c = XBeeClient()
+mgr = GameManager()
+c = XBeeClient(mgr, '/dev/tty.usbserial-DN040E8Y', 9600)
 c.listen_loop()
